ExactCalculator: Add haptic button widget
This is a basic extension of Button that provides haptic feedback when
the user touches the button (*not* when they release, to reduce latency).
It uses the system's predefined virtual key haptic feedback effect for
modern vibration effects that the user is already familiar with and
automatic response to changes in system vibration settings.